+++ frontend/src/App.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { Route, Routes } from 'react-router-dom';
-import LoginSignup from './components/LoginSignup/LoginSignup';
-import Home from './components/Home/Home';
 import { ConversationChannelPage } from './pages/ConversationChannelPage';
 import { ConversationPage } from './pages/ConversationPage';
-import User from './components/User/User';
+import LoginSignup from './components/LoginSignup/LoginSignup';
+import Home from './components/Home/Home';
+import User from './pages/UserPage';
 
 function App() {
   return (
